+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2021-02-21
+
+### Added
+
+- Printer support (experimental)
+- Background with some tips
+
+### Changed
+
+- Automatically start X session
+- Update ISO builder image to `aitorpazos/create-debian-iso:0.2.0`
+
 ## [0.2.0] - 2021-02-14
 
 ### Added